+++ test/integration/GCClassic/integrationTest.sh
@@ -142,8 +142,12 @@ fi
 thisDir=$(pwd -P)
 
 # Load common functions
-. "${thisDir}/commonFunctionsForTests.sh"
-
+if [[ -f "../../shared/commonFunctionsForTests.sh" ]]; then
+    . "${thisDir}/../../shared/commonFunctionsForTests.sh"
+elif [[ -f "${thisDir}/commonFunctionsForTests.sh" ]]; then
+    . "${thisDir}/commonFunctionsForTests.sh"
+fi
+    
 #=============================================================================
 # Create integration test directories in the root folder
 #=============================================================================
@@ -185,7 +189,7 @@ if [[ "x${scheduler}" == "xSLURM" ]]; then
     # Remove LSF #BSUB tags
     sed_ie '/#BSUB -q REQUESTED_PARTITION/d' "${scriptsDir}/integrationTestCompile.sh"
     sed_ie '/#BSUB -n 8/d'                   "${scriptsDir}/integrationTestCompile.sh"
-    sed_ie '/#BSUB -W 00:30/d'               "${scriptsDir}/integrationTestCompile.sh"
+    sed_ie '/#BSUB -W 0:30/d'                "${scriptsDir}/integrationTestCompile.sh"
     sed_ie '/#BSUB -o lsf-%J.txt/d'          "${scriptsDir}/integrationTestCompile.sh"
     sed_ie \
 	'/#BSUB -R "rusage\[mem=8GB\] span\[ptile=1\] select\[mem < 1TB\]"/d' \
@@ -195,7 +199,7 @@ if [[ "x${scheduler}" == "xSLURM" ]]; then
 	"${scriptsDir}/integrationTestCompile.sh"
     sed_ie '/#BSUB -q REQUESTED_PARTITION/d' "${scriptsDir}/integrationTestExecute.sh"
     sed_ie '/#BSUB -n 24/d'                  "${scriptsDir}/integrationTestExecute.sh"
-    sed_ie '/#BSUB -W 2:00/d'                "${scriptsDir}/integrationTestExecute.sh"
+    sed_ie '/#BSUB -W 6:00/d'                "${scriptsDir}/integrationTestExecute.sh"
     sed_ie '/#BSUB -o lsf-%J.txt/d'          "${scriptsDir}/integrationTestExecute.sh"
     sed_ie \
 	'/#BSUB -R "rusage\[mem=90GB\] span\[ptile=1\] select\[mem < 2TB\]"/d' \
@@ -232,14 +236,14 @@ elif [[ "x${scheduler}" == "xLSF" ]]; then
     # Remove SLURM #SBATCH tags
     sed_ie '/#SBATCH -c 8/d'                   "${scriptsDir}/integrationTestCompile.sh"
     sed_ie '/#SBATCH -N 1/d'                   "${scriptsDir}/integrationTestCompile.sh"
-    sed_ie '/#SBATCH -t 0-00:30/d'             "${scriptsDir}/integrationTestCompile.sh"
+    sed_ie '/#SBATCH -t 0-0:30/d'              "${scriptsDir}/integrationTestCompile.sh"
     sed_ie '/#SBATCH -p REQUESTED_PARTITION/d' "${scriptsDir}/integrationTestCompile.sh"
     sed_ie '/#SBATCH --mem=8000/d'             "${scriptsDir}/integrationTestCompile.sh"
     sed_ie '/#SBATCH -p REQUESTED_PARTITION/d' "${scriptsDir}/integrationTestCompile.sh"
     sed_ie '/#SBATCH --mail-type=END/d'        "${scriptsDir}/integrationTestCompile.sh"
     sed_ie '/#SBATCH -c 24/d'                  "${scriptsDir}/integrationTestExecute.sh"
     sed_ie '/#SBATCH -N 1/d'                   "${scriptsDir}/integrationTestExecute.sh"
-    sed_ie '/#SBATCH -t 0-2:00/d'              "${scriptsDir}/integrationTestExecute.sh"
+    sed_ie '/#SBATCH -t 0-6:00/d'              "${scriptsDir}/integrationTestExecute.sh"
     sed_ie '/#SBATCH -p REQUESTED_PARTITION/d' "${scriptsDir}/integrationTestExecute.sh"
     sed_ie '/#SBATCH --mem=90000/d'            "${scriptsDir}/integrationTestExecute.sh"
     sed_ie '/#SBATCH --mail-type=END/d'        "${scriptsDir}/integrationTestExecute.sh"