+++ test/integration/GCHP/integrationTestCreate.sh
@@ -69,7 +69,8 @@ head_hco=$(export GIT_DISCOVERY_ACROSS_FILESYSTEM=1; \
 	   git -C "${hemcoDir}" log --oneline --no-decorate -1)
 
 # Source the script containing utility functions and variables
-. "${geosChemDir}/test/shared/commonFunctionsForTests.sh"
+commonFuncs="${geosChemDir}/test/shared/commonFunctionsForTests.sh"
+. "${commonFuncs}"
 
 # Echo header
 printf "${SEP_MAJOR}\n"
@@ -138,11 +139,11 @@ ln -s "${superProjectDir}" ${itRoot}/CodeDir
 #=============================================================================
 
 printf "\nCopying run scripts to: ${itRoot}/${SCRIPTS_DIR}\n"
-cp -f ${envFile}                            ${envDir}/gchp.env
-cp -f ${thisDir}/integrationTest*.sh        ${scriptsDir}
-cp -f ${thisDir}/commonFunctionsForTests.sh ${scriptsDir}
-cp -f ${thisDir}/README.md                  ${scriptsDir}
-cp -f ${thisDir}/README.testroot.md         ${itRoot}/README.md
+cp -f ${envFile}                     ${envDir}/gchp.env
+cp -f ${thisDir}/integration*.sh     ${scriptsDir}
+cp -f ${commonFuncs}                 ${scriptsDir}
+cp -f ${thisDir}/README.md           ${scriptsDir}
+cp -f ${thisDir}/README.testroot.md  ${itRoot}/README.md
 
 # Log file with echoback from rundir creation
 log="${logsDir}/createIntegrationTests.log"
@@ -190,6 +191,7 @@ cd "${thisDir}"
 # Free local variables
 unset binDir
 unset buildDir
+unset commonFuncs
 unset dir
 unset envDir
 unset geosChemDir